feat(qubes): manage devices attached to sys-usb
In Qubes, you need to shutdown a qube before you can manipulate its
devices. It may become a problem when you don't have an integrated
keyboard and/or touchpad, because after shutting down sys-usb you lose
the ability to interact with your system (if you didn't setup ssh, but
we talk here about dom0, so you actually don't want to do this). So you
need at least some kind of a scheduled start for sys-usb. This is what
this script does, along with adding/removing a specified device in
between the shutdown and the start of sys-usb.