feat(qubes): manipulate disposable VM volume
In Qubes, at least on my PC hardware, changing volume in dom0 does
nothing for other qubes. I believe a proper setup of sys-audio is
required to make it work across all of them. But for now I don't need
all the qubes - I'm usually only interested in the qube that is used
for browsing, and chances are it's a disposable. Moreover, most of the
time there's only one, so there is a simple logic to avoid typing the
random digits assigned to it as its name.